fix test + restructure
